move: Token 관련 코드를 `jwt` 도메인 하위로 이전
